requirements.txt
# JobGuard AI — Model Training
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
nltk>=3.8
joblib>=1.2.0
scipy>=1.9.0
matplotlib>=3.5.0
seaborn>=0.12.0
wordcloud>=1.9.0
fastapi>=0.110.0
uvicorn>=0.30.0
httpx>=0.27.0
PyYAML>=6.0
nbformat>=5.9.0
pytest>=8.0.0

# Optional: for downloading extra datasets
datasets>=2.14.0

# Optional: for XGBoost (better accuracy)
xgboost>=1.7.0
